requirements.txt
# === Data Fetching ===
yfinance>=0.2.31
fredapi>=0.5.1
requests>=2.31.0

# === Data Processing ===
pandas>=2.1.0
numpy>=1.25.0
scipy>=1.11.0

# === Local LLM ===
ollama>=0.3.0

# === Database ===
sqlalchemy>=2.0.0
alembic>=1.12.0

# === Scheduling ===
apscheduler>=3.10.0

# === Dashboard ===
streamlit>=1.28.0
plotly>=5.17.0

# === Config ===
pyyaml>=6.0.1
python-dotenv>=1.0.0

# === Alerts ===
slack-sdk>=3.23.0

# === SEC Filing Parsing ===
beautifulsoup4>=4.12.0
lxml>=4.9.0

# === Testing ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
